email-summary-agent/requirements.txt
fastmcp>=3.2.2
fastapi
uvicorn
httpx
azure-identity
python-dotenv
openai
